notebook/ngan/unzip/requirements.txt
fastapi
uvicorn[standard]
google-cloud-storage
google-auth
